Move GlobFilter into a separate header/unit
Addresses a long standing todo in workerd where we duplicate
GlobFilter there.
