+++ c++/src/kj/BUILD.bazel
@@ -13,6 +13,7 @@ cc_library(
         "encoding.c++",
         "exception.c++",
         "filesystem.c++",
+        "glob-filter.c++",
         "hash.c++",
         "io.c++",
         "list.c++",
@@ -42,6 +43,7 @@ cc_library(
         "exception.h",
         "filesystem.h",
         "function.h",
+        "glob-filter.h",
         "hash.h",
         "io.h",
         "list.h",
