+++ c++/src/kj/test.c++
@@ -68,95 +68,6 @@ size_t TestCase::iterCount() {
 
 // =======================================================================================
 
-namespace _ {  // private
-
-GlobFilter::GlobFilter(const char* pattern): pattern(heapString(pattern)) {}
-GlobFilter::GlobFilter(ArrayPtr<const char> pattern): pattern(heapString(pattern)) {}
-
-bool GlobFilter::matches(StringPtr name) {
-  // Get out your computer science books. We're implementing a non-deterministic finite automaton.
-  //
-  // Our NDFA has one "state" corresponding to each character in the pattern.
-  //
-  // As you may recall, an NDFA can be transformed into a DFA where every state in the DFA
-  // represents some combination of states in the NDFA. Therefore, we actually have to store a
-  // list of states here. (Actually, what we really want is a set of states, but because our
-  // patterns are mostly non-cyclic a list of states should work fine and be a bit more efficient.)
-
-  // Our state list starts out pointing only at the start of the pattern.
-  states.resize(0);
-  states.add(0);
-
-  Vector<uint> scratch;
-
-  // Iterate through each character in the name.
-  for (char c: name) {
-    // Pull the current set of states off to the side, so that we can populate `states` with the
-    // new set of states.
-    Vector<uint> oldStates = kj::mv(states);
-    states = kj::mv(scratch);
-    states.resize(0);
-
-    // The pattern can omit a leading path. So if we're at a '/' then enter the state machine at
-    // the beginning on the next char.
-    if (c == '/' || c == '\\') {
-      states.add(0);
-    }
-
-    // Process each state.
-    for (uint state: oldStates) {
-      applyState(c, state);
-    }
-
-    // Store the previous state vector for reuse.
-    scratch = kj::mv(oldStates);
-  }
-
-  // If any one state is at the end of the pattern (or at a wildcard just before the end of the
-  // pattern), we have a match.
-  for (uint state: states) {
-    while (state < pattern.size() && pattern[state] == '*') {
-      ++state;
-    }
-    if (state == pattern.size()) {
-      return true;
-    }
-  }
-  return false;
-}
-
-void GlobFilter::applyState(char c, int state) {
-  if (state < pattern.size()) {
-    switch (pattern[state]) {
-      case '*':
-        // At a '*', we both re-add the current state and attempt to match the *next* state.
-        if (c != '/' && c != '\\') {  // '*' doesn't match '/'.
-          states.add(state);
-        }
-        applyState(c, state + 1);
-        break;
-
-      case '?':
-        // A '?' matches one character (never a '/').
-        if (c != '/' && c != '\\') {
-          states.add(state + 1);
-        }
-        break;
-
-      default:
-        // Any other character matches only itself.
-        if (c == pattern[state]) {
-          states.add(state + 1);
-        }
-        break;
-    }
-  }
-}
-
-}  // namespace _ (private)
-
-// =======================================================================================
-
 namespace {
 
 class TestExceptionCallback: public ExceptionCallback {
@@ -255,7 +166,7 @@ public:
       }
     }
 
-    _::GlobFilter filter(filePattern);
+    kj::GlobFilter filter(filePattern);
 
     for (TestCase* testCase = testCasesHead; testCase != nullptr; testCase = testCase->next) {
       if (!testCase->matchedFilter && filter.matches(testCase->file) &&
